重构Bat

+++ install-docker.bat
@@ -19,7 +19,7 @@
 @set IMAGE_NAME=skeleton-spring-cloud
 
 @rem 删除target，有时候mvn会clean失败，需事先强制删除target
-rmdir /s/q %PROJECT_NAME%\target
+if exist %PROJECT_NAME%\target rmdir /s/q %PROJECT_NAME%\target
 
 @rem 执行相关模块的install
 call mvn clean install -DskipTests -pl skeleton-engine,%PROJECT_NAME% -am
